Start working on activity Feed

+++ base/views.py
@@ -1,9 +1,13 @@
 from django.http import HttpResponse
 from django.db.models import Q
 from django.shortcuts import render, redirect
-
+from django.contrib.auth.models import User
+from django.contrib import messages
 from base.forms import RoomForm
-from base.models import Room, Topic
+from django.contrib.auth import authenticate, login, logout
+from base.models import Room, Topic, Message
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 # rooms = [
 #     {'id': 1, 'name': 'Lets learn python'},
@@ -11,6 +15,56 @@ from base.models import Room, Topic
 #     {'id': 3, 'name': 'Lets learn Java'},
 #     {'id': 4, 'name': 'Lets learn JavaScript'},
 # ]
+
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Invalid username or password')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid username or password')
+
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid username or password')
+
+    context = { 'form': form }
+
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
 
@@ -24,18 +78,34 @@ def home(request):
 
     topics = Topic.objects.all()
 
-    context = { 'rooms': enumerate(rooms, start=1), 'topics': topics }
+    room_count = rooms.count()
+
+    context = { 'rooms': enumerate(rooms, start=1), 'topics': topics, 'room_count': room_count}
 
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
     
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
 
-    context = {'room': room}
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body'),
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
 
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
 
     form = RoomForm()
@@ -50,8 +120,12 @@ def createRoom(request):
 
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your not allowed here')
 
     form = RoomForm(instance=room)
 
@@ -64,12 +138,30 @@ def updateRoom(request, pk):
     context = {'form': form }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
 
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your not allowed here')
 
     if request.method == 'POST':
         room.delete()
         return redirect('home')
 
     return render(request, 'base/delete.html', {'obj': room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('Your not allowed here')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+
+    return render(request, 'base/delete.html', {'obj': message})
